Network.Control

Description

Common parts to control network protocols.

Constants for flow control.

defaultMaxStreams :: Int

Default max streams. (64)

defaultMaxStreamData :: Int

Default max data of a stream. (256K bytes)

defaultMaxData :: Int

Default max data of a connection. (1M bytes)

Flow control for sending

data TxFlow

Flow for sending

Constructors

TxFlow

Fields

txfSent :: Int

txfLimit :: Int

newTxFlow :: WindowSize -> TxFlow

Creating TX flow with an initial window size.

txWindowSize :: TxFlow -> WindowSize

txfLimit - txfSent.

type WindowSize = Int

Window size.

Flow control for receiving

data RxFlow

Flow for receiving

Constructors

RxFlow

Fields

rxfWindow :: WindowSize

rxfConsumed :: Int

rxfReceived :: Int

rxfLimit :: Int

newRxFlow :: WindowSize -> RxFlow

Creating RX flow with an initial window size.

data FlowControlType

The representation of window size update.

Constructors

FCTWindowUpdate
HTTP/2 style

FCTMaxData
QUIC style

maybeOpenRxWindow

Arguments

:: Int
The consumed size.

-> FlowControlType

-> RxFlow

-> (RxFlow, Maybe Int)
Just if the size should be informed to the peer.

When an application consumed received data, this function should be called to update rxfConsumed. If the window size is less than the half of the initial window. the representation of window size update is returned.

checkRxLimit

Arguments

:: Int
The size of received data.

-> RxFlow

-> (RxFlow, Bool)
Acceptable if True.

Checking if received data is acceptable against the current window.

LRU cache

data LRUCache k v

Sized cache based on least recently used.

empty

Arguments

:: Int
The size of LRUCache.

-> LRUCache k v

Empty LRUCache.

insert :: Ord k => k -> v -> LRUCache k v -> LRUCache k v

Inserting.

delete :: Ord k => k -> LRUCache k v -> LRUCache k v

Deleting.

lookup :: Ord k => k -> LRUCache k v -> Maybe v

Looking up.

Rate control

data Rate

Type for rating.

newRate :: IO Rate

Creating a new Rate.

getRate :: Rate -> IO Int

Getting the current rate. If one or more seconds have passed since the previous call, the counter is re-initialized with 1 and it is returned. Otherwise, incremented counter number is returned.

addRate :: Rate -> Int -> IO Int

Getting the current rate. If one or more seconds have passed since the previous call, the counter is re-initialized with the second argument and it is returned. Otherwise, increased counter number is returned.